変更 - クエリを作成するロジックをファイルに切り出す
